CONFDEV-59478 Allowing for map data sources keyed by propertyName as well
If an alias is present but no value is found for the alias in the source datamap
check if the propertyName is mapped. This happens in braid when we do manual mapping
between datasources since we coerce the data key to be a static string instead of
building it dynamically based on the user's query.